changed files for check functions
